requirements.txt
alembic==1.7.3
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.25
Werkzeug==2.0.1
